+++ colour/models/rgb/transfer_functions/__init__.py
