+++ colour/models/rgb/transfer_functions/alexa_log_c.py
