+++ colour/models/rgb/transfer_functions/bt_1886.py
