+++ colour/models/rgb/transfer_functions/canon_clog.py
