+++ colour/models/rgb/transfer_functions/cineon.py
