+++ colour/models/rgb/transfer_functions/gamma.py
