+++ colour/models/rgb/transfer_functions/linear.py
