+++ colour/models/rgb/transfer_functions/panalog.py
