+++ colour/models/rgb/transfer_functions/pivoted_log.py
