+++ colour/models/rgb/transfer_functions/sony_slog.py
