+++ colour/models/rgb/transfer_functions/srgb.py
