+++ colour/models/rgb/transfer_functions/tests/__init__.py
