+++ colour/models/rgb/transfer_functions/tests/tests_bt_2020.py
@@ -2,7 +2,7 @@
 # -*- coding: utf-8 -*-
 
 """
-Defines unit tests for :mod:`colour.models.rgb.conversion_functions.sony_slog`
+Defines unit tests for :mod:`colour.models.rgb.transfer_functions.bt_2020`
 module.
 """
 
@@ -11,9 +11,7 @@ from __future__ import division, unicode_literals
 import numpy as np
 import unittest
 
-from colour.models.rgb.conversion_functions import (
-    log_encoding_SLog,
-    log_decoding_SLog)
+from colour.models.rgb.transfer_functions import oecf_BT2020, eocf_BT2020
 from colour.utilities import ignore_numpy_errors
 
 __author__ = 'Colour Developers'
@@ -23,151 +21,151 @@ __maintainer__ = 'Colour Developers'
 __email__ = 'colour-science@googlegroups.com'
 __status__ = 'Production'
 
-__all__ = ['TestLogEncoding_SLog',
-           'TestLogDecoding_SLog']
+__all__ = ['TestOecf_BT2020',
+           'TestEocf_BT2020']
 
 
-class TestLogEncoding_SLog(unittest.TestCase):
+class TestOecf_BT2020(unittest.TestCase):
     """
-    Defines :func:`colour.models.rgb.conversion_functions.sony_slog.\
-log_encoding_SLog` definition unit tests methods.
+    Defines :func:`colour.models.rgb.transfer_functions.bt_2020.oecf_BT2020`
+    definition unit tests methods.
     """
 
-    def test_log_encoding_SLog(self):
+    def test_oecf_BT2020(self):
         """
-        Tests :func:`colour.models.rgb.conversion_functions.sony_slog.\
-log_encoding_SLog` definition.
+        Tests :func:`colour.models.rgb.transfer_functions.bt_2020.\
+oecf_BT2020` definition.
         """
 
         self.assertAlmostEqual(
-            log_encoding_SLog(0.0),
-            0.030001222851889303,
+            oecf_BT2020(0.0),
+            0.0,
             places=7)
 
         self.assertAlmostEqual(
-            log_encoding_SLog(0.18),
-            0.35998784642215442,
+            oecf_BT2020(0.18),
+            0.4090077288641504,
             places=7)
 
         self.assertAlmostEqual(
-            log_encoding_SLog(1.0),
-            0.65352925122530825,
+            oecf_BT2020(1.0),
+            1.0,
             places=7)
 
-    def test_n_dimensional_log_encoding_SLog(self):
+    def test_n_dimensional_oecf_BT2020(self):
         """
-        Tests :func:`colour.models.rgb.conversion_functions.sony_slog.\
-log_encoding_SLog` definition n-dimensional arrays support.
+        Tests :func:`colour.models.rgb.transfer_functions.bt_2020.\
+oecf_BT2020` definition n-dimensional arrays support.
         """
 
         L = 0.18
-        V = 0.35998784642215442
+        V = 0.4090077288641504
         np.testing.assert_almost_equal(
-            log_encoding_SLog(L),
+            oecf_BT2020(L),
             V,
             decimal=7)
 
         L = np.tile(L, 6)
         V = np.tile(V, 6)
         np.testing.assert_almost_equal(
-            log_encoding_SLog(L),
+            oecf_BT2020(L),
             V,
             decimal=7)
 
         L = np.reshape(L, (2, 3))
         V = np.reshape(V, (2, 3))
         np.testing.assert_almost_equal(
-            log_encoding_SLog(L),
+            oecf_BT2020(L),
             V,
             decimal=7)
 
         L = np.reshape(L, (2, 3, 1))
         V = np.reshape(V, (2, 3, 1))
         np.testing.assert_almost_equal(
-            log_encoding_SLog(L),
+            oecf_BT2020(L),
             V,
             decimal=7)
 
     @ignore_numpy_errors
-    def test_nan_log_encoding_SLog(self):
+    def test_nan_oecf_BT2020(self):
         """
-        Tests :func:`colour.models.rgb.conversion_functions.sony_slog.\
-log_encoding_SLog` definition nan support.
+        Tests :func:`colour.models.rgb.transfer_functions.bt_2020.\
+oecf_BT2020` definition nan support.
         """
 
-        log_encoding_SLog(
+        oecf_BT2020(
             np.array([-1.0, 0.0, 1.0, -np.inf, np.inf, np.nan]))
 
 
-class TestLogDecoding_SLog(unittest.TestCase):
+class TestEocf_BT2020(unittest.TestCase):
     """
-    Defines :func:`colour.models.rgb.conversion_functions.sony_slog.\
-log_decoding_SLog` definition unit tests methods.
+    Defines :func:`colour.models.rgb.transfer_functions.bt_2020.eocf_BT2020`
+    definition unit tests methods.
     """
 
-    def test_log_decoding_SLog(self):
+    def test_eocf_BT2020(self):
         """
-        Tests :func:`colour.models.rgb.conversion_functions.sony_slog.\
-log_decoding_SLog` definition.
+        Tests :func:`colour.models.rgb.transfer_functions.bt_2020.\
+eocf_BT2020` definition.
         """
 
         self.assertAlmostEqual(
-            log_decoding_SLog(0.0),
-            -0.0055458144460765538,
+            eocf_BT2020(0.0),
+            0.0,
             places=7)
 
         self.assertAlmostEqual(
-            log_decoding_SLog(0.35998784642215442),
+            eocf_BT2020(0.4090077288641504),
             0.18,
             places=7)
 
         self.assertAlmostEqual(
-            log_decoding_SLog(0.65352925122530825),
+            eocf_BT2020(1.0),
             1.0,
             places=7)
 
-    def test_n_dimensional_log_decoding_SLog(self):
+    def test_n_dimensional_eocf_BT2020(self):
         """
-        Tests :func:`colour.models.rgb.conversion_functions.sony_slog.\
-log_decoding_SLog` definition n-dimensional arrays support.
+        Tests :func:`colour.models.rgb.transfer_functions.bt_2020.\
+eocf_BT2020` definition n-dimensional arrays support.
         """
 
-        V = 0.35998784642215442
+        V = 0.4090077288641504
         L = 0.18
         np.testing.assert_almost_equal(
-            log_decoding_SLog(V),
+            eocf_BT2020(V),
             L,
             decimal=7)
 
         V = np.tile(V, 6)
         L = np.tile(L, 6)
         np.testing.assert_almost_equal(
-            log_decoding_SLog(V),
+            eocf_BT2020(V),
             L,
             decimal=7)
 
         V = np.reshape(V, (2, 3))
         L = np.reshape(L, (2, 3))
         np.testing.assert_almost_equal(
-            log_decoding_SLog(V),
+            eocf_BT2020(V),
             L,
             decimal=7)
 
         V = np.reshape(V, (2, 3, 1))
         L = np.reshape(L, (2, 3, 1))
         np.testing.assert_almost_equal(
-            log_decoding_SLog(V),
+            eocf_BT2020(V),
             L,
             decimal=7)
 
     @ignore_numpy_errors
-    def test_nan_log_decoding_SLog(self):
+    def test_nan_eocf_BT2020(self):
         """
-        Tests :func:`colour.models.rgb.conversion_functions.sony_slog.\
-log_decoding_SLog` definition nan support.
+        Tests :func:`colour.models.rgb.transfer_functions.bt_2020.\
+eocf_BT2020` definition nan support.
         """
 
-        log_decoding_SLog(
+        eocf_BT2020(
             np.array([-1.0, 0.0, 1.0, -np.inf, np.inf, np.nan]))
