+++ colour/models/rgb/transfer_functions/tests/tests_bt_709.py
@@ -2,7 +2,7 @@
 # -*- coding: utf-8 -*-
 
 """
-Defines unit tests for :mod:`colour.models.rgb.conversion_functions.bt_2020`
+Defines unit tests for :mod:`colour.models.rgb.transfer_functions.bt_709`
 module.
 """
 
@@ -11,7 +11,7 @@ from __future__ import division, unicode_literals
 import numpy as np
 import unittest
 
-from colour.models.rgb.conversion_functions import oecf_BT2020, eocf_BT2020
+from colour.models.rgb.transfer_functions import oecf_BT709, eocf_BT709
 from colour.utilities import ignore_numpy_errors
 
 __author__ = 'Colour Developers'
@@ -21,151 +21,151 @@ __maintainer__ = 'Colour Developers'
 __email__ = 'colour-science@googlegroups.com'
 __status__ = 'Production'
 
-__all__ = ['TestOecf_BT2020',
-           'TestEocf_BT2020']
+__all__ = ['TestOecf_BT709',
+           'TestEocf_BT709']
 
 
-class TestOecf_BT2020(unittest.TestCase):
+class TestOecf_BT709(unittest.TestCase):
     """
-    Defines :func:`colour.models.rgb.conversion_functions.bt_2020.oecf_BT2020`
+    Defines :func:`colour.models.rgb.transfer_functions.bt_709.oecf_BT709`
     definition unit tests methods.
     """
 
-    def test_oecf_BT2020(self):
+    def test_oecf_BT709(self):
         """
-        Tests :func:`colour.models.rgb.conversion_functions.bt_2020.\
-oecf_BT2020` definition.
+        Tests :func:`colour.models.rgb.transfer_functions.bt_709.\
+oecf_BT709` definition.
         """
 
         self.assertAlmostEqual(
-            oecf_BT2020(0.0),
+            oecf_BT709(0.0),
             0.0,
             places=7)
 
         self.assertAlmostEqual(
-            oecf_BT2020(0.18),
+            oecf_BT709(0.18),
             0.4090077288641504,
             places=7)
 
         self.assertAlmostEqual(
-            oecf_BT2020(1.0),
+            oecf_BT709(1.0),
             1.0,
             places=7)
 
-    def test_n_dimensional_oecf_BT2020(self):
+    def test_n_dimensional_oecf_BT709(self):
         """
-        Tests :func:`colour.models.rgb.conversion_functions.bt_2020.\
-oecf_BT2020` definition n-dimensional arrays support.
+        Tests :func:`colour.models.rgb.transfer_functions.bt_709.\
+oecf_BT709` definition n-dimensional arrays support.
         """
 
         L = 0.18
         V = 0.4090077288641504
         np.testing.assert_almost_equal(
-            oecf_BT2020(L),
+            oecf_BT709(L),
             V,
             decimal=7)
 
         L = np.tile(L, 6)
         V = np.tile(V, 6)
         np.testing.assert_almost_equal(
-            oecf_BT2020(L),
+            oecf_BT709(L),
             V,
             decimal=7)
 
         L = np.reshape(L, (2, 3))
         V = np.reshape(V, (2, 3))
         np.testing.assert_almost_equal(
-            oecf_BT2020(L),
+            oecf_BT709(L),
             V,
             decimal=7)
 
         L = np.reshape(L, (2, 3, 1))
         V = np.reshape(V, (2, 3, 1))
         np.testing.assert_almost_equal(
-            oecf_BT2020(L),
+            oecf_BT709(L),
             V,
             decimal=7)
 
     @ignore_numpy_errors
-    def test_nan_oecf_BT2020(self):
+    def test_nan_oecf_BT709(self):
         """
-        Tests :func:`colour.models.rgb.conversion_functions.bt_2020.\
-oecf_BT2020` definition nan support.
+        Tests :func:`colour.models.rgb.transfer_functions.bt_709.\
+oecf_BT709` definition nan support.
         """
 
-        oecf_BT2020(
+        oecf_BT709(
             np.array([-1.0, 0.0, 1.0, -np.inf, np.inf, np.nan]))
 
 
-class TestEocf_BT2020(unittest.TestCase):
+class TestEocf_BT709(unittest.TestCase):
     """
-    Defines :func:`colour.models.rgb.conversion_functions.bt_2020.eocf_BT2020`
+    Defines :func:`colour.models.rgb.transfer_functions.bt_709.eocf_BT709`
     definition unit tests methods.
     """
 
-    def test_eocf_BT2020(self):
+    def test_eocf_BT709(self):
         """
-        Tests :func:`colour.models.rgb.conversion_functions.bt_2020.\
-eocf_BT2020` definition.
+        Tests :func:`colour.models.rgb.transfer_functions.bt_709.\
+eocf_BT709` definition.
         """
 
         self.assertAlmostEqual(
-            eocf_BT2020(0.0),
+            eocf_BT709(0.0),
             0.0,
             places=7)
 
         self.assertAlmostEqual(
-            eocf_BT2020(0.4090077288641504),
+            eocf_BT709(0.4090077288641504),
             0.18,
             places=7)
 
         self.assertAlmostEqual(
-            eocf_BT2020(1.0),
+            eocf_BT709(1.0),
             1.0,
             places=7)
 
-    def test_n_dimensional_eocf_BT2020(self):
+    def test_n_dimensional_eocf_BT709(self):
         """
-        Tests :func:`colour.models.rgb.conversion_functions.bt_2020.\
-eocf_BT2020` definition n-dimensional arrays support.
+        Tests :func:`colour.models.rgb.transfer_functions.bt_709.\
+eocf_BT709` definition n-dimensional arrays support.
         """
 
         V = 0.4090077288641504
         L = 0.18
         np.testing.assert_almost_equal(
-            eocf_BT2020(V),
+            eocf_BT709(V),
             L,
             decimal=7)
 
         V = np.tile(V, 6)
         L = np.tile(L, 6)
         np.testing.assert_almost_equal(
-            eocf_BT2020(V),
+            eocf_BT709(V),
             L,
             decimal=7)
 
         V = np.reshape(V, (2, 3))
         L = np.reshape(L, (2, 3))
         np.testing.assert_almost_equal(
-            eocf_BT2020(V),
+            eocf_BT709(V),
             L,
             decimal=7)
 
         V = np.reshape(V, (2, 3, 1))
         L = np.reshape(L, (2, 3, 1))
         np.testing.assert_almost_equal(
-            eocf_BT2020(V),
+            eocf_BT709(V),
             L,
             decimal=7)
 
     @ignore_numpy_errors
-    def test_nan_eocf_BT2020(self):
+    def test_nan_eocf_BT709(self):
         """
-        Tests :func:`colour.models.rgb.conversion_functions.bt_2020.\
-eocf_BT2020` definition nan support.
+        Tests :func:`colour.models.rgb.transfer_functions.bt_709.\
+eocf_BT709` definition nan support.
         """
 
-        eocf_BT2020(
+        eocf_BT709(
             np.array([-1.0, 0.0, 1.0, -np.inf, np.inf, np.nan]))
