+++ colour/models/rgb/transfer_functions/tests/tests_canon_clog.py
@@ -2,8 +2,8 @@
 # -*- coding: utf-8 -*-
 
 """
-Defines unit tests for :mod:`colour.models.rgb.conversion_functions.\
-panasonic_vlog` module.
+Defines unit tests for :mod:`colour.models.rgb.transfer_functions.\
+canon_clog` module.
 """
 
 from __future__ import division, unicode_literals
@@ -11,9 +11,9 @@ from __future__ import division, unicode_literals
 import numpy as np
 import unittest
 
-from colour.models.rgb.conversion_functions import (
-    log_encoding_VLog,
-    log_decoding_VLog)
+from colour.models.rgb.transfer_functions import (
+    log_encoding_CLog,
+    log_decoding_CLog)
 from colour.utilities import ignore_numpy_errors
 
 __author__ = 'Colour Developers'
@@ -23,151 +23,151 @@ __maintainer__ = 'Colour Developers'
 __email__ = 'colour-science@googlegroups.com'
 __status__ = 'Production'
 
-__all__ = ['TestLogEncoding_VLog',
-           'TestLogDecoding_VLog']
+__all__ = ['TestLogEncoding_CLog',
+           'TestLogDecoding_CLog']
 
 
-class TestLogEncoding_VLog(unittest.TestCase):
+class TestLogEncoding_CLog(unittest.TestCase):
     """
-    Defines :func:`colour.models.rgb.conversion_functions.panasonic_vlog.\
-log_encoding_VLog` definition unit tests methods.
+    Defines :func:`colour.models.rgb.transfer_functions.canon_clog.\
+log_encoding_CLog` definition unit tests methods.
     """
 
-    def test_log_encoding_VLog(self):
+    def test_log_encoding_CLog(self):
         """
-        Tests :func:`colour.models.rgb.conversion_functions.panasonic_vlog.\
-log_encoding_VLog` definition.
+        Tests :func:`colour.models.rgb.transfer_functions.canon_clog.\
+log_encoding_CLog` definition.
         """
 
         self.assertAlmostEqual(
-            log_encoding_VLog(0.0),
-            0.125,
+            log_encoding_CLog(0.0),
+            0.073059700000000005,
             places=7)
 
         self.assertAlmostEqual(
-            log_encoding_VLog(0.18),
-            0.42331144876013616,
+            log_encoding_CLog(0.18),
+            0.31201285555039493,
             places=7)
 
         self.assertAlmostEqual(
-            log_encoding_VLog(1.0),
-            0.5991177001581459,
+            log_encoding_CLog(1.0),
+            0.62740830453765284,
             places=7)
 
-    def test_n_dimensional_log_encoding_VLog(self):
+    def test_n_dimensional_log_encoding_CLog(self):
         """
-        Tests :func:`colour.models.rgb.conversion_functions.panasonic_vlog.\
-log_encoding_VLog` definition n-dimensional arrays support.
+        Tests :func:`colour.models.rgb.transfer_functions.canon_clog.\
+log_encoding_CLog` definition n-dimensional arrays support.
         """
 
         L = 0.18
-        V = 0.42331144876013616
+        V = 0.31201285555039493
         np.testing.assert_almost_equal(
-            log_encoding_VLog(L),
+            log_encoding_CLog(L),
             V,
             decimal=7)
 
         L = np.tile(L, 6)
         V = np.tile(V, 6)
         np.testing.assert_almost_equal(
-            log_encoding_VLog(L),
+            log_encoding_CLog(L),
             V,
             decimal=7)
 
         L = np.reshape(L, (2, 3))
         V = np.reshape(V, (2, 3))
         np.testing.assert_almost_equal(
-            log_encoding_VLog(L),
+            log_encoding_CLog(L),
             V,
             decimal=7)
 
         L = np.reshape(L, (2, 3, 1))
         V = np.reshape(V, (2, 3, 1))
         np.testing.assert_almost_equal(
-            log_encoding_VLog(L),
+            log_encoding_CLog(L),
             V,
             decimal=7)
 
     @ignore_numpy_errors
-    def test_nan_log_encoding_VLog(self):
+    def test_nan_log_encoding_CLog(self):
         """
-        Tests :func:`colour.models.rgb.conversion_functions.panasonic_vlog.\
-log_encoding_VLog` definition nan support.
+        Tests :func:`colour.models.rgb.transfer_functions.canon_clog.\
+log_encoding_CLog` definition nan support.
         """
 
-        log_encoding_VLog(
+        log_encoding_CLog(
             np.array([-1.0, 0.0, 1.0, -np.inf, np.inf, np.nan]))
 
 
-class TestLogDecoding_VLog(unittest.TestCase):
+class TestLogDecoding_CLog(unittest.TestCase):
     """
-    Defines :func:`colour.models.rgb.conversion_functions.panasonic_vlog.\
-log_decoding_VLog` definition unit tests methods.
+    Defines :func:`colour.models.rgb.transfer_functions.canon_clog.\
+log_decoding_CLog` definition unit tests methods.
     """
 
-    def test_log_decoding_VLog(self):
+    def test_log_decoding_CLog(self):
         """
-        Tests :func:`colour.models.rgb.conversion_functions.panasonic_vlog.\
-log_decoding_VLog` definition.
+        Tests :func:`colour.models.rgb.transfer_functions.canon_clog.\
+log_decoding_CLog` definition.
         """
 
         self.assertAlmostEqual(
-            log_decoding_VLog(0.125),
+            log_decoding_CLog(0.073059700000000005),
             0.0,
             places=7)
 
         self.assertAlmostEqual(
-            log_decoding_VLog(0.42331144876013616),
+            log_decoding_CLog(0.31201285555039493),
             0.18,
             places=7)
 
         self.assertAlmostEqual(
-            log_decoding_VLog(0.5991177001581459),
+            log_decoding_CLog(0.62740830453765284),
             1.0,
             places=7)
 
-    def test_n_dimensional_log_decoding_VLog(self):
+    def test_n_dimensional_log_decoding_CLog(self):
         """
-        Tests :func:`colour.models.rgb.conversion_functions.panasonic_vlog.\
-log_decoding_VLog` definition n-dimensional arrays support.
+        Tests :func:`colour.models.rgb.transfer_functions.canon_clog.\
+log_decoding_CLog` definition n-dimensional arrays support.
         """
 
-        V = 0.42331144876013616
+        V = 0.31201285555039493
         L = 0.18
         np.testing.assert_almost_equal(
-            log_decoding_VLog(V),
+            log_decoding_CLog(V),
             L,
             decimal=7)
 
         V = np.tile(V, 6)
         L = np.tile(L, 6)
         np.testing.assert_almost_equal(
-            log_decoding_VLog(V),
+            log_decoding_CLog(V),
             L,
             decimal=7)
 
         V = np.reshape(V, (2, 3))
         L = np.reshape(L, (2, 3))
         np.testing.assert_almost_equal(
-            log_decoding_VLog(V),
+            log_decoding_CLog(V),
             L,
             decimal=7)
 
         V = np.reshape(V, (2, 3, 1))
         L = np.reshape(L, (2, 3, 1))
         np.testing.assert_almost_equal(
-            log_decoding_VLog(V),
+            log_decoding_CLog(V),
             L,
             decimal=7)
 
     @ignore_numpy_errors
-    def test_nan_log_decoding_VLog(self):
+    def test_nan_log_decoding_CLog(self):
         """
-        Tests :func:`colour.models.rgb.conversion_functions.panasonic_vlog.\
-log_decoding_VLog` definition nan support.
+        Tests :func:`colour.models.rgb.transfer_functions.canon_clog.\
+log_decoding_CLog` definition nan support.
         """
 
-        log_decoding_VLog(
+        log_decoding_CLog(
             np.array([-1.0, 0.0, 1.0, -np.inf, np.inf, np.nan]))
