+++ colour/models/rgb/transfer_functions/tests/tests_linear.py
@@ -2,7 +2,7 @@
 # -*- coding: utf-8 -*-
 
 """
-Defines unit tests for :mod:`colour.models.rgb.conversion_functions.linear`
+Defines unit tests for :mod:`colour.models.rgb.transfer_functions.linear`
 module.
 """
 
@@ -11,7 +11,7 @@ from __future__ import division, unicode_literals
 import numpy as np
 import unittest
 
-from colour.models.rgb.conversion_functions import linear_function
+from colour.models.rgb.transfer_functions import linear_function
 from colour.utilities import ignore_numpy_errors
 
 __author__ = 'Colour Developers'
@@ -26,13 +26,13 @@ __all__ = ['TestLinearFunction']
 
 class TestLinearFunction(unittest.TestCase):
     """
-    Defines :func:`colour.models.rgb.conversion_functions.linear.\
+    Defines :func:`colour.models.rgb.transfer_functions.linear.\
 linear_function` definition unit tests methods.
     """
 
     def test_linear_function(self):
         """
-        Tests :func:`colour.models.rgb.conversion_functions.linear.\
+        Tests :func:`colour.models.rgb.transfer_functions.linear.\
 linear_function` definition.
         """
 
@@ -44,7 +44,7 @@ linear_function` definition.
 
     def test_n_dimensional_linear_function(self):
         """
-        Tests :func:`colour.models.rgb.conversion_functions.linear.\
+        Tests :func:`colour.models.rgb.transfer_functions.linear.\
 linear_function` definition n-dimensional arrays support.
         """
 
@@ -79,7 +79,7 @@ linear_function` definition n-dimensional arrays support.
     @ignore_numpy_errors
     def test_nan_linear_function(self):
         """
-        Tests :func:`colour.models.rgb.conversion_functions.linear.\
+        Tests :func:`colour.models.rgb.transfer_functions.linear.\
 linear_function` definition nan support.
         """
