+++ colour/models/rgb/transfer_functions/tests/tests_panasonic_vlog.py
@@ -2,8 +2,8 @@
 # -*- coding: utf-8 -*-
 
 """
-Defines unit tests for :mod:`colour.models.rgb.conversion_functions.dci_p3`
-module.
+Defines unit tests for :mod:`colour.models.rgb.transfer_functions.\
+panasonic_vlog` module.
 """
 
 from __future__ import division, unicode_literals
@@ -11,9 +11,9 @@ from __future__ import division, unicode_literals
 import numpy as np
 import unittest
 
-from colour.models.rgb.conversion_functions import (
-    log_encoding_DCIP3,
-    log_decoding_DCIP3)
+from colour.models.rgb.transfer_functions import (
+    log_encoding_VLog,
+    log_decoding_VLog)
 from colour.utilities import ignore_numpy_errors
 
 __author__ = 'Colour Developers'
@@ -23,151 +23,151 @@ __maintainer__ = 'Colour Developers'
 __email__ = 'colour-science@googlegroups.com'
 __status__ = 'Production'
 
-__all__ = ['TestLogEncoding_DCIP3',
-           'TestLogDecoding_DCIP3']
+__all__ = ['TestLogEncoding_VLog',
+           'TestLogDecoding_VLog']
 
 
-class TestLogEncoding_DCIP3(unittest.TestCase):
+class TestLogEncoding_VLog(unittest.TestCase):
     """
-    Defines :func:`colour.models.rgb.conversion_functions.dci_p3.\
-log_encoding_DCIP3` definition unit tests methods.
+    Defines :func:`colour.models.rgb.transfer_functions.panasonic_vlog.\
+log_encoding_VLog` definition unit tests methods.
     """
 
-    def test_log_encoding_DCIP3(self):
+    def test_log_encoding_VLog(self):
         """
-        Tests :func:`colour.models.rgb.conversion_functions.dci_p3.\
-log_encoding_DCIP3` definition.
+        Tests :func:`colour.models.rgb.transfer_functions.panasonic_vlog.\
+log_encoding_VLog` definition.
         """
 
         self.assertAlmostEqual(
-            log_encoding_DCIP3(0.0),
-            0.0,
+            log_encoding_VLog(0.0),
+            0.125,
             places=7)
 
         self.assertAlmostEqual(
-            log_encoding_DCIP3(0.18),
-            461.99220597484737,
+            log_encoding_VLog(0.18),
+            0.42331144876013616,
             places=7)
 
         self.assertAlmostEqual(
-            log_encoding_DCIP3(1.0),
-            893.44598340527841,
+            log_encoding_VLog(1.0),
+            0.5991177001581459,
             places=7)
 
-    def test_n_dimensional_log_encoding_DCIP3(self):
+    def test_n_dimensional_log_encoding_VLog(self):
         """
-        Tests :func:`colour.models.rgb.conversion_functions.dci_p3.\
-log_encoding_DCIP3` definition n-dimensional arrays support.
+        Tests :func:`colour.models.rgb.transfer_functions.panasonic_vlog.\
+log_encoding_VLog` definition n-dimensional arrays support.
         """
 
         L = 0.18
-        V = 461.99220597484737
+        V = 0.42331144876013616
         np.testing.assert_almost_equal(
-            log_encoding_DCIP3(L),
+            log_encoding_VLog(L),
             V,
             decimal=7)
 
         L = np.tile(L, 6)
         V = np.tile(V, 6)
         np.testing.assert_almost_equal(
-            log_encoding_DCIP3(L),
+            log_encoding_VLog(L),
             V,
             decimal=7)
 
         L = np.reshape(L, (2, 3))
         V = np.reshape(V, (2, 3))
         np.testing.assert_almost_equal(
-            log_encoding_DCIP3(L),
+            log_encoding_VLog(L),
             V,
             decimal=7)
 
         L = np.reshape(L, (2, 3, 1))
         V = np.reshape(V, (2, 3, 1))
         np.testing.assert_almost_equal(
-            log_encoding_DCIP3(L),
+            log_encoding_VLog(L),
             V,
             decimal=7)
 
     @ignore_numpy_errors
-    def test_nan_log_encoding_DCIP3(self):
+    def test_nan_log_encoding_VLog(self):
         """
-        Tests :func:`colour.models.rgb.conversion_functions.dci_p3.\
-log_encoding_DCIP3` definition nan support.
+        Tests :func:`colour.models.rgb.transfer_functions.panasonic_vlog.\
+log_encoding_VLog` definition nan support.
         """
 
-        log_encoding_DCIP3(
+        log_encoding_VLog(
             np.array([-1.0, 0.0, 1.0, -np.inf, np.inf, np.nan]))
 
 
-class TestLogDecoding_DCIP3(unittest.TestCase):
+class TestLogDecoding_VLog(unittest.TestCase):
     """
-    Defines :func:`colour.models.rgb.conversion_functions.dci_p3.
-log_decoding_DCIP3` definition unit tests methods.
+    Defines :func:`colour.models.rgb.transfer_functions.panasonic_vlog.\
+log_decoding_VLog` definition unit tests methods.
     """
 
-    def test_log_decoding_DCIP3(self):
+    def test_log_decoding_VLog(self):
         """
-        Tests :func:`colour.models.rgb.conversion_functions.dci_p3.\
-log_decoding_DCIP3` definition.
+        Tests :func:`colour.models.rgb.transfer_functions.panasonic_vlog.\
+log_decoding_VLog` definition.
         """
 
         self.assertAlmostEqual(
-            log_decoding_DCIP3(0.0),
+            log_decoding_VLog(0.125),
             0.0,
             places=7)
 
         self.assertAlmostEqual(
-            log_decoding_DCIP3(461.99220597484737),
+            log_decoding_VLog(0.42331144876013616),
             0.18,
             places=7)
 
         self.assertAlmostEqual(
-            log_decoding_DCIP3(893.44598340527841),
+            log_decoding_VLog(0.5991177001581459),
             1.0,
             places=7)
 
-    def test_n_dimensional_log_decoding_DCIP3(self):
+    def test_n_dimensional_log_decoding_VLog(self):
         """
-        Tests :func:`colour.models.rgb.conversion_functions.dci_p3.\
-log_decoding_DCIP3` definition n-dimensional arrays support.
+        Tests :func:`colour.models.rgb.transfer_functions.panasonic_vlog.\
+log_decoding_VLog` definition n-dimensional arrays support.
         """
 
-        V = 461.99220597484737
+        V = 0.42331144876013616
         L = 0.18
         np.testing.assert_almost_equal(
-            log_decoding_DCIP3(V),
+            log_decoding_VLog(V),
             L,
             decimal=7)
 
         V = np.tile(V, 6)
         L = np.tile(L, 6)
         np.testing.assert_almost_equal(
-            log_decoding_DCIP3(V),
+            log_decoding_VLog(V),
             L,
             decimal=7)
 
         V = np.reshape(V, (2, 3))
         L = np.reshape(L, (2, 3))
         np.testing.assert_almost_equal(
-            log_decoding_DCIP3(V),
+            log_decoding_VLog(V),
             L,
             decimal=7)
 
         V = np.reshape(V, (2, 3, 1))
         L = np.reshape(L, (2, 3, 1))
         np.testing.assert_almost_equal(
-            log_decoding_DCIP3(V),
+            log_decoding_VLog(V),
             L,
             decimal=7)
 
     @ignore_numpy_errors
-    def test_nan_log_decoding_DCIP3(self):
+    def test_nan_log_decoding_VLog(self):
         """
-        Tests :func:`colour.models.rgb.conversion_functions.dci_p3.\
-log_decoding_DCIP3` definition nan support.
+        Tests :func:`colour.models.rgb.transfer_functions.panasonic_vlog.\
+log_decoding_VLog` definition nan support.
         """
 
-        log_decoding_DCIP3(
+        log_decoding_VLog(
             np.array([-1.0, 0.0, 1.0, -np.inf, np.inf, np.nan]))
