+++ colour/models/rgb/transfer_functions/tests/tests_srgb.py
@@ -2,8 +2,8 @@
 # -*- coding: utf-8 -*-
 
 """
-Defines unit tests for :mod:`colour.models.rgb.conversion_functions.\
-canon_clog` module.
+Defines unit tests for :mod:`colour.models.rgb.transfer_functions.sRGB`
+module.
 """
 
 from __future__ import division, unicode_literals
@@ -11,9 +11,7 @@ from __future__ import division, unicode_literals
 import numpy as np
 import unittest
 
-from colour.models.rgb.conversion_functions import (
-    log_encoding_CLog,
-    log_decoding_CLog)
+from colour.models.rgb.transfer_functions import oecf_sRGB, eocf_sRGB
 from colour.utilities import ignore_numpy_errors
 
 __author__ = 'Colour Developers'
@@ -23,151 +21,151 @@ __maintainer__ = 'Colour Developers'
 __email__ = 'colour-science@googlegroups.com'
 __status__ = 'Production'
 
-__all__ = ['TestLogEncoding_CLog',
-           'TestLogDecoding_CLog']
+__all__ = ['TestOecf_sRGB',
+           'TestEocf_sRGB']
 
 
-class TestLogEncoding_CLog(unittest.TestCase):
+class TestOecf_sRGB(unittest.TestCase):
     """
-    Defines :func:`colour.models.rgb.conversion_functions.canon_clog.\
-log_encoding_CLog` definition unit tests methods.
+    Defines :func:`colour.models.rgb.transfer_functions.sRGB.oecf_sRGB`
+    definition unit tests methods.
     """
 
-    def test_log_encoding_CLog(self):
+    def test_oecf_sRGB(self):
         """
-        Tests :func:`colour.models.rgb.conversion_functions.canon_clog.\
-log_encoding_CLog` definition.
+        Tests :func:`colour.models.rgb.transfer_functions.sRGB.\
+oecf_sRGB` definition.
         """
 
         self.assertAlmostEqual(
-            log_encoding_CLog(0.0),
-            0.073059700000000005,
-            places=7)
-
-        self.assertAlmostEqual(
-            log_encoding_CLog(0.18),
-            0.31201285555039493,
-            places=7)
-
-        self.assertAlmostEqual(
-            log_encoding_CLog(1.0),
-            0.62740830453765284,
-            places=7)
-
-    def test_n_dimensional_log_encoding_CLog(self):
-        """
-        Tests :func:`colour.models.rgb.conversion_functions.canon_clog.\
-log_encoding_CLog` definition n-dimensional arrays support.
-        """
-
-        L = 0.18
-        V = 0.31201285555039493
-        np.testing.assert_almost_equal(
-            log_encoding_CLog(L),
-            V,
-            decimal=7)
-
-        L = np.tile(L, 6)
-        V = np.tile(V, 6)
-        np.testing.assert_almost_equal(
-            log_encoding_CLog(L),
-            V,
-            decimal=7)
-
-        L = np.reshape(L, (2, 3))
-        V = np.reshape(V, (2, 3))
-        np.testing.assert_almost_equal(
-            log_encoding_CLog(L),
-            V,
-            decimal=7)
-
-        L = np.reshape(L, (2, 3, 1))
-        V = np.reshape(V, (2, 3, 1))
-        np.testing.assert_almost_equal(
-            log_encoding_CLog(L),
-            V,
-            decimal=7)
-
-    @ignore_numpy_errors
-    def test_nan_log_encoding_CLog(self):
-        """
-        Tests :func:`colour.models.rgb.conversion_functions.canon_clog.\
-log_encoding_CLog` definition nan support.
-        """
-
-        log_encoding_CLog(
-            np.array([-1.0, 0.0, 1.0, -np.inf, np.inf, np.nan]))
-
-
-class TestLogDecoding_CLog(unittest.TestCase):
-    """
-    Defines :func:`colour.models.rgb.conversion_functions.canon_clog.\
-log_decoding_CLog` definition unit tests methods.
-    """
-
-    def test_log_decoding_CLog(self):
-        """
-        Tests :func:`colour.models.rgb.conversion_functions.canon_clog.\
-log_decoding_CLog` definition.
-        """
-
-        self.assertAlmostEqual(
-            log_decoding_CLog(0.073059700000000005),
+            oecf_sRGB(0.0),
             0.0,
             places=7)
 
         self.assertAlmostEqual(
-            log_decoding_CLog(0.31201285555039493),
+            oecf_sRGB(0.18),
+            0.46135612950044164,
+            places=7)
+
+        self.assertAlmostEqual(
+            oecf_sRGB(1.0),
+            1.0,
+            places=7)
+
+    def test_n_dimensional_oecf_sRGB(self):
+        """
+        Tests :func:`colour.models.rgb.transfer_functions.sRGB.\
+oecf_sRGB` definition n-dimensional arrays support.
+        """
+
+        L = 0.18
+        V = 0.46135612950044164
+        np.testing.assert_almost_equal(
+            oecf_sRGB(L),
+            V,
+            decimal=7)
+
+        L = np.tile(L, 6)
+        V = np.tile(V, 6)
+        np.testing.assert_almost_equal(
+            oecf_sRGB(L),
+            V,
+            decimal=7)
+
+        L = np.reshape(L, (2, 3))
+        V = np.reshape(V, (2, 3))
+        np.testing.assert_almost_equal(
+            oecf_sRGB(L),
+            V,
+            decimal=7)
+
+        L = np.reshape(L, (2, 3, 1))
+        V = np.reshape(V, (2, 3, 1))
+        np.testing.assert_almost_equal(
+            oecf_sRGB(L),
+            V,
+            decimal=7)
+
+    @ignore_numpy_errors
+    def test_nan_oecf_sRGB(self):
+        """
+        Tests :func:`colour.models.rgb.transfer_functions.sRGB.\
+oecf_sRGB` definition nan support.
+        """
+
+        oecf_sRGB(
+            np.array([-1.0, 0.0, 1.0, -np.inf, np.inf, np.nan]))
+
+
+class TestEocf_sRGB(unittest.TestCase):
+    """
+    Defines :func:`colour.models.rgb.transfer_functions.sRGB.eocf_sRGB`
+    definition unit tests methods.
+    """
+
+    def test_eocf_sRGB(self):
+        """
+        Tests :func:`colour.models.rgb.transfer_functions.sRGB.\
+eocf_sRGB` definition.
+        """
+
+        self.assertAlmostEqual(
+            eocf_sRGB(0.0),
+            0.0,
+            places=7)
+
+        self.assertAlmostEqual(
+            eocf_sRGB(0.46135612950044164),
             0.18,
             places=7)
 
         self.assertAlmostEqual(
-            log_decoding_CLog(0.62740830453765284),
+            eocf_sRGB(1.0),
             1.0,
             places=7)
 
-    def test_n_dimensional_log_decoding_CLog(self):
+    def test_n_dimensional_eocf_sRGB(self):
         """
-        Tests :func:`colour.models.rgb.conversion_functions.canon_clog.\
-log_decoding_CLog` definition n-dimensional arrays support.
+        Tests :func:`colour.models.rgb.transfer_functions.sRGB.\
+eocf_sRGB` definition n-dimensional arrays support.
         """
 
-        V = 0.31201285555039493
+        V = 0.46135612950044164
         L = 0.18
         np.testing.assert_almost_equal(
-            log_decoding_CLog(V),
+            eocf_sRGB(V),
             L,
             decimal=7)
 
         V = np.tile(V, 6)
         L = np.tile(L, 6)
         np.testing.assert_almost_equal(
-            log_decoding_CLog(V),
+            eocf_sRGB(V),
             L,
             decimal=7)
 
         V = np.reshape(V, (2, 3))
         L = np.reshape(L, (2, 3))
         np.testing.assert_almost_equal(
-            log_decoding_CLog(V),
+            eocf_sRGB(V),
             L,
             decimal=7)
 
         V = np.reshape(V, (2, 3, 1))
         L = np.reshape(L, (2, 3, 1))
         np.testing.assert_almost_equal(
-            log_decoding_CLog(V),
+            eocf_sRGB(V),
             L,
             decimal=7)
 
     @ignore_numpy_errors
-    def test_nan_log_decoding_CLog(self):
+    def test_nan_eocf_sRGB(self):
         """
-        Tests :func:`colour.models.rgb.conversion_functions.canon_clog.\
-log_decoding_CLog` definition nan support.
+        Tests :func:`colour.models.rgb.transfer_functions.sRGB.\
+eocf_sRGB` definition nan support.
         """
 
-        log_decoding_CLog(
+        eocf_sRGB(
             np.array([-1.0, 0.0, 1.0, -np.inf, np.inf, np.nan]))
