+++ colour/models/rgb/transfer_functions/viper_log.py
